iwlwav-iw: Align nl80211_copy.h with nl80211.h in driver
As part of the move from 4.16 to 4.19 kernel, there are changes
made in the upstream to nl80211.h. The upstream changes are pulled
into the driver during the backport process to support kernel 4.19.

Align nl80211_copy.h to ensure all upstream changes pulled into
driver to nl80211.h is also part of the hostap maintained copy.